requirements.txt
google-api-python-client>=2.100.0
youtube-transcript-api>=0.6.1
openai>=1.3.0
python-telegram-bot>=20.6
python-dotenv>=1.0.0
requests>=2.31.0
langchain>=0.1.0
langchain-openai>=0.1.0
PyYAML>=6.0
